The delivery classifiers get their own file, and a claim names its gate
classify.go crossed the 500-line cap at 502. The three classifiers that moved
share a concept rather than a line count: a bounce, an undelivered message and
the pipes all tell a reader something did not happen instead of handing them
something to do, and that difference decides both the row's words and the
screen it lands on.

ownerFrom's claim that three lanes answer a zero alike now names the two tests
that hold it.
