+++ backend/internal/compose/attention/classify.go
@@ -177,7 +177,11 @@ func classifyFailedApproval(item crmcontracts.AttentionItem, asOf time.Time) ran
 	row := base(item, levelPromise, "system", "you_believe_it_happened")
 	row.Because = []crmcontracts.WorklistReason{reason("approved_and_failed", nil)}
 	return ranked{
-		ownerRef: ownedByWhoeverIsReading(), item: row, occurredAt: occurredOf(item, asOf)}
+		item:       row,
+		occurredAt: occurredOf(item, asOf),
+		// Carried back to the person who APPROVED it, by a lane bound to them.
+		ownerRef: ownedByWhoeverIsReading(),
+	}
 }
 
 // classifyIntroduction: a colleague is waiting on this reader to answer.
@@ -439,45 +443,4 @@ func classifyTask(item crmcontracts.AttentionItem, asOf time.Time) ranked {
 		// is a row a reader cannot make sense of.
 		ownerRef: ownerFromAssignee(item.AssigneeId),
 	}
-}
-
-// classifyBounce: a customer never received something the rep believes they
-// sent. A consequence with a named customer and a verb, so it ranks as its own
-// row rather than disappearing into an aggregate.
-func classifyBounce(item crmcontracts.AttentionItem, asOf time.Time) ranked {
-	row := base(item, levelPromise, "system", "customer_never_received")
-	row.Because = []crmcontracts.WorklistReason{reason("approved_and_failed", nil)}
-	return ranked{
-		ownerRef: ownedByWhoeverIsReading(), item: row, occurredAt: occurredOf(item, asOf)}
-}
-
-// classifyUndelivered: a message the rep believes they sent and which never
-// left. The belief is the damage — they are waiting on a reply to something
-// nobody has — so it sits with the broken promises rather than with the system
-// news, exactly where a bounce sits.
-//
-// It is a separate consequence from the bounce beside it: nobody received this
-// one because nobody was ever sent it, and the reader's move is to send it
-// rather than to fix an address.
-func classifyUndelivered(item crmcontracts.AttentionItem, asOf time.Time) ranked {
-	row := base(item, levelPromise, "system", "you_believe_it_happened")
-	row.Because = []crmcontracts.WorklistReason{reason("approved_and_failed", nil)}
-	return ranked{
-		ownerRef: ownedByWhoeverIsReading(), item: row, occurredAt: occurredOf(item, asOf)}
-}
-
-// classifySystem: the pipes. A mailbox that stopped capturing makes every quiet
-// claim on this page suspect, which is why it is here at all rather than only in
-// an admin screen.
-func classifySystem(item crmcontracts.AttentionItem, asOf time.Time) ranked {
-	consequence := crmcontracts.WorklistItemConsequence("work_blocked")
-	if item.Source == "capture_health" || item.Source == "sync_health" {
-		consequence = "mailbox_blind"
-	}
-	if item.Source == "notice" {
-		consequence = valueNone
-	}
-	row := base(item, levelBlocking, "system", consequence)
-	return ranked{
-		ownerRef: ownedByWhoeverIsReading(), item: row, occurredAt: occurredOf(item, asOf)}
 }
